Move CredentialsBindingHandler extension point to extensions plugin
This way, plugins will be able to just depend on
pipeline-model-extensions if they want to implement a handler. I also
moved the actual implementations out of CredentialsBindingHandler and
into a better package - I couldn't move CredentialsBindingHandler
itself into a better package because I didn't want to break compatibility.
