+++ pipeline-model-extensions/src/main/java/org/jenkinsci/plugins/pipeline/modeldefinition/model/CredentialsBindingHandler.java
@@ -1,7 +1,7 @@
 /*
  * The MIT License
  *
- * Copyright (c) 2016, CloudBees, Inc.
+ * Copyright (c) 2018, CloudBees, Inc.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -20,7 +20,6 @@
  * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
- *
  */
 
 package org.jenkinsci.plugins.pipeline.modeldefinition.model;
@@ -28,28 +27,17 @@ package org.jenkinsci.plugins.pipeline.modeldefinition.model;
 import com.cloudbees.plugins.credentials.CredentialsProvider;
 import com.cloudbees.plugins.credentials.common.IdCredentials;
 import com.cloudbees.plugins.credentials.common.StandardCredentials;
-import com.cloudbees.plugins.credentials.common.StandardUsernamePasswordCredentials;
-import hudson.Extension;
 import hudson.ExtensionList;
 import hudson.ExtensionPoint;
 import hudson.model.Run;
 import org.apache.commons.lang.StringUtils;
 import org.jenkinsci.plugins.credentialsbinding.MultiBinding;
 import org.jenkinsci.plugins.credentialsbinding.impl.CredentialNotFoundException;
-import org.jenkinsci.plugins.credentialsbinding.impl.FileBinding;
-import org.jenkinsci.plugins.credentialsbinding.impl.StringBinding;
-import org.jenkinsci.plugins.credentialsbinding.impl.UsernamePasswordBinding;
-import org.jenkinsci.plugins.credentialsbinding.impl.UsernamePasswordMultiBinding;
-import org.jenkinsci.plugins.plaincredentials.FileCredentials;
-import org.jenkinsci.plugins.plaincredentials.StringCredentials;
 
 import javax.annotation.CheckForNull;
 import javax.annotation.Nonnull;
 import java.io.Serializable;
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
@@ -87,7 +75,7 @@ public abstract class CredentialsBindingHandler<C extends StandardCredentials> i
     @Nonnull
     public static Set<Class<? extends StandardCredentials>> supportedTypes() {
         Set<Class<? extends StandardCredentials>> set = new HashSet<>();
-        for (CredentialsBindingHandler handler : all()) {
+        for (CredentialsBindingHandler<?> handler : all()) {
             set.add(handler.type());
         }
         return set;
@@ -153,91 +141,4 @@ public abstract class CredentialsBindingHandler<C extends StandardCredentials> i
         }
     }
 
-    @Extension
-    public static class UsernamePasswordHandler extends CredentialsBindingHandler<StandardUsernamePasswordCredentials> {
-        @Nonnull
-        @Override
-        public List<MultiBinding<StandardUsernamePasswordCredentials>> toBindings(String varName, String credentialsId) {
-            List<MultiBinding<StandardUsernamePasswordCredentials>> bindings = new ArrayList<>();
-            bindings.add(new UsernamePasswordBinding(varName, credentialsId));
-            bindings.add(new UsernamePasswordMultiBinding(varName + "_USR",
-                    varName + "_PSW",
-                    credentialsId));
-            return bindings;
-        }
-
-        @Nonnull
-        @Override
-        public Class<? extends StandardCredentials> type() {
-            return StandardUsernamePasswordCredentials.class;
-        }
-
-        @Nonnull
-        @Override
-        public List<Map<String, Object>> getWithCredentialsParameters(String credentialsId) {
-            Map<String, Object> map = new HashMap<>();
-            map.put("$class", UsernamePasswordBinding.class.getName());
-            map.put("variable", new EnvVarResolver());
-            map.put("credentialsId", credentialsId);
-            Map<String, Object> map2 = new HashMap<>();
-            map2.put("$class", UsernamePasswordMultiBinding.class.getName());
-            map2.put("usernameVariable", new EnvVarResolver("%s_USR"));
-            map2.put("passwordVariable", new EnvVarResolver("%s_PSW"));
-            map2.put("credentialsId", credentialsId);
-            return Arrays.asList(map, map2);
-        }
-
-    }
-
-    @Extension
-    public static class FileCredentialsHandler extends CredentialsBindingHandler<FileCredentials> {
-
-        @Nonnull
-        @Override
-        public List<MultiBinding<FileCredentials>> toBindings(String varName, String credentialsId) {
-            return Collections.<MultiBinding<FileCredentials>>singletonList(new FileBinding(varName, credentialsId));
-        }
-
-        @Nonnull
-        @Override
-        public Class<? extends StandardCredentials> type() {
-            return FileCredentials.class;
-        }
-
-        @Nonnull
-        @Override
-        public List<Map<String, Object>> getWithCredentialsParameters(String credentialsId) {
-            Map<String, Object> map = new HashMap<>();
-            map.put("$class", FileBinding.class.getName());
-            map.put("variable", new EnvVarResolver());
-            map.put("credentialsId", credentialsId);
-            return Collections.singletonList(map);
-        }
-    }
-
-    @Extension
-    public static class StringCredentialsHandler extends CredentialsBindingHandler<StringCredentials> {
-
-        @Nonnull
-        @Override
-        public List<MultiBinding<StringCredentials>> toBindings(String varName, String credentialsId) {
-            return Collections.<MultiBinding<StringCredentials>>singletonList(new StringBinding(varName, credentialsId));
-        }
-
-        @Nonnull
-        @Override
-        public Class<? extends StandardCredentials> type() {
-            return StringCredentials.class;
-        }
-
-        @Nonnull
-        @Override
-        public List<Map<String, Object>> getWithCredentialsParameters(String credentialsId) {
-            Map<String, Object> map = new HashMap<>();
-            map.put("$class", StringBinding.class.getName());
-            map.put("variable", new EnvVarResolver());
-            map.put("credentialsId", credentialsId);
-            return Collections.singletonList(map);
-        }
-    }
 }
